+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## UNRELEASED
+
+- [#927](https://github.com/crypto-org-chain/chain-main/pull/927) Integrate versiondb.
+
 *February 20, 2023*
 
 ## v4.2.3